+++ src/main.cpp
@@ -36,14 +36,6 @@
 /* Platformio libraries */
 #include <ArduinoJson.h>
 #include <time.h>
-#include <wolfssl/options.h>
-#include <wolfssl/ssl.h>
-#include <wolfssl/wolfcrypt/rsa.h>
-#include <wolfssl/wolfcrypt/sha256.h>
-#include <wolfssl/wolfcrypt/signature.h>
-#include <wolfssl/wolfcrypt/hash.h>
-#include <wolfssl/wolfcrypt/asn.h>
-#include <wolfssl/wolfcrypt/ed25519.h>
 
 /* Project libraries */
 #include "BriandTorEsp32Config.hxx"
@@ -52,6 +44,7 @@
 #include "BriandNet.hxx"
 #include "BriandTorRelay.hxx"
 #include "BriandTorCircuit.hxx"
+#include "BriandTorCertificateUtils.hxx"
 
 using namespace std;
 
@@ -121,13 +114,38 @@ void setup() {
         string key = "1234567890123456";
         int t = 0;
         auto config = make_unique<Briand::BriandTorEsp32Config>(key);
-        Serial.printf("[TEST] AES Encryption test, string <%s>\n", test.c_str());
+        Serial.printf("[TEST] AES Encryption test, string <%s> with key <%s>\n", test.c_str(), key.c_str());
         auto buf = config->encrypt(test);
         Serial.print("[TEST] Encrypted Bytes: ");
-        for (int i = 0; i<test.length(); i++)
-           Serial.printf("%0x ", buf.get()[i] );
+		Briand::BriandUtils::PrintOldStyleByteBuffer(buf.get(), test.length(), test.length()+1, test.length());
         Serial.printf("\n[TEST] Decrypted Bytes: <%s>\n", config->decrypt(buf, test.length()).c_str());
         Serial.println("[TEST] AES Test success.");
+		buf.reset();
+		config.reset();
+
+		// Test SHA256
+		string testMessage = string("546F7220544C53205253412F456432353531392063726F73732D63657274696669636174651EAE084E96C9150FAE941A28DD7A9B718EFD0F759D7021A9754A717C65D19B350006EA89");
+		string expResult = string("457E063D5CE929FE98AF745D1DA20306422E9203298E69408F75B0595EA703C7");
+		auto message = Briand::BriandUtils::HexStringToVector(testMessage, "");
+		Serial.printf("[TEST] Perform SHA256 hash of:  %s\n", testMessage.c_str());
+		Serial.printf("[TEST] Expected output:         %s\n", expResult.c_str());
+		auto hash = Briand::BriandTorCertificateUtils::GetDigest_SHA256(message);
+		Serial.printf("[TEST] SHA256 computed hash is: ");
+		Briand::BriandUtils::PrintByteBuffer(*(hash.get()), hash->size()+1, hash->size());
+		auto expResultV = Briand::BriandUtils::HexStringToVector(expResult, "");
+
+		if (expResultV->size() != hash->size()) Serial.printf("[TEST] FAIL SHA256, sizes do not math (%d against expected %d).\n", hash->size(), expResultV->size());
+		else {
+			bool differentFound = false;
+			for (int i=0; i<hash->size() && !differentFound; i++)
+				differentFound = ( hash->at(i) != expResultV->at(i) );
+			if (!differentFound) Serial.printf("[TEST] SHA256 test success!\n");
+			else Serial.printf("[TEST] SHA256 test failure! (hash does not match expected result).\n");
+		}
+
+		message.reset();
+		hash.reset();
+		expResultV.reset();
     }
 
 	// Init WiFi to AP+STA 
@@ -538,6 +556,31 @@ void executeCommand(string& cmd) {
     }
 	else if (cmd.compare("synctime") == 0) {
         syncTimeWithNTP();
+    }
+	else if (cmd.compare("sha") == 0) {
+        // Test SHA256
+		string testMessage = string("546F7220544C53205253412F456432353531392063726F73732D63657274696669636174651EAE084E96C9150FAE941A28DD7A9B718EFD0F759D7021A9754A717C65D19B350006EA89");
+		string expResult = string("457E063D5CE929FE98AF745D1DA20306422E9203298E69408F75B0595EA703C7");
+		auto message = Briand::BriandUtils::HexStringToVector(testMessage, "");
+		Serial.printf("[TEST] Perform SHA256 hash of:  %s\n", testMessage.c_str());
+		Serial.printf("[TEST] Expected output:         %s\n", expResult.c_str());
+		auto hash = Briand::BriandTorCertificateUtils::GetDigest_SHA256(message);
+		Serial.printf("[TEST] SHA256 computed hash is: ");
+		Briand::BriandUtils::PrintByteBuffer(*(hash.get()), hash->size()+1, hash->size());
+		auto expResultV = Briand::BriandUtils::HexStringToVector(expResult, "");
+
+		if (expResultV->size() != hash->size()) Serial.printf("[TEST] FAIL SHA256, sizes do not math (%d against expected %d).\n", hash->size(), expResultV->size());
+		else {
+			bool differentFound = false;
+			for (int i=0; i<hash->size() && !differentFound; i++)
+				differentFound = ( hash->at(i) != expResultV->at(i) );
+			if (!differentFound) Serial.printf("[TEST] SHA256 test success!\n");
+			else Serial.printf("[TEST] SHA256 test failure! (hash does not match expected result).\n");
+		}
+
+		message.reset();
+		hash.reset();
+		expResultV.reset();
     }
     else if (cmd.compare("devinfo") == 0) {
         Serial.printf("CPU Frequency: %uMHz\n", ESP.getCpuFreqMHz());